+++ src/v0/bitcoin/map/output.rs
@@ -13,27 +13,27 @@ use crate::v0::bitcoin::map::Map;
 use crate::v0::bitcoin::{raw, Error};
 
 /// Type: Redeem ScriptBuf PSBT_OUT_REDEEM_SCRIPT = 0x00
-const PSBT_OUT_REDEEM_SCRIPT: u8 = 0x00;
+const PSBT_OUT_REDEEM_SCRIPT: u64 = 0x00;
 /// Type: Witness ScriptBuf PSBT_OUT_WITNESS_SCRIPT = 0x01
-const PSBT_OUT_WITNESS_SCRIPT: u8 = 0x01;
+const PSBT_OUT_WITNESS_SCRIPT: u64 = 0x01;
 /// Type: BIP 32 Derivation Path PSBT_OUT_BIP32_DERIVATION = 0x02
-const PSBT_OUT_BIP32_DERIVATION: u8 = 0x02;
+const PSBT_OUT_BIP32_DERIVATION: u64 = 0x02;
 /// Type: Taproot Internal Key PSBT_OUT_TAP_INTERNAL_KEY = 0x05
-const PSBT_OUT_TAP_INTERNAL_KEY: u8 = 0x05;
+const PSBT_OUT_TAP_INTERNAL_KEY: u64 = 0x05;
 /// Type: Taproot Tree PSBT_OUT_TAP_TREE = 0x06
-const PSBT_OUT_TAP_TREE: u8 = 0x06;
+const PSBT_OUT_TAP_TREE: u64 = 0x06;
 /// Type: Taproot Key BIP 32 Derivation Path PSBT_OUT_TAP_BIP32_DERIVATION = 0x07
-const PSBT_OUT_TAP_BIP32_DERIVATION: u8 = 0x07;
+const PSBT_OUT_TAP_BIP32_DERIVATION: u64 = 0x07;
 /// Type: Output Amount PSBT_OUT_AMOUNT = 0x03
-const PSBT_OUT_AMOUNT: u8 = 0x03;
+const PSBT_OUT_AMOUNT: u64 = 0x03;
 /// Type: Output Script PSBT_OUT_SCRIPT = 0x04
-const PSBT_OUT_SCRIPT: u8 = 0x04;
+const PSBT_OUT_SCRIPT: u64 = 0x04;
 /// Type: Silent Payment recipient data PSBT_OUT_SP_V0_INFO = 0x09
-const PSBT_OUT_SP_V0_INFO: u8 = 0x09;
+const PSBT_OUT_SP_V0_INFO: u64 = 0x09;
 /// Type: Silent Payment Label used PSBT_OUT_SP_V0_LABEL = 0x0a
-const PSBT_OUT_SP_V0_LABEL: u8 = 0x0a;
+const PSBT_OUT_SP_V0_LABEL: u64 = 0x0a;
 /// Type: Proprietary Use Type PSBT_IN_PROPRIETARY = 0xFC
-const PSBT_OUT_PROPRIETARY: u8 = 0xFC;
+const PSBT_OUT_PROPRIETARY: u64 = 0xFC;
 
 /// A key-value map for an output of the corresponding index in the unsigned
 /// transaction.
